Make UMAP callback pickleable (#6402)
The structure of this callback object should really be cleaned up (by moving the callback to the stack before the call to UMAP we can avoid having an object with pointers to itself _and_ better manage the lifetime of the callback instance). This would also let us avoid implementing this callback in cython, which mucks up the default pickling implementation. As is though this is the easiest way to get things pickleable. Users will still need to define `__reduce__` themselves if they add any state to an instance, but given that this code was added for an internal use case and hasn't seen any changes or expansion in 6 years I don't suspect there are many external users of this API. Can always clean this up later if a user asks about it.

Fixes #5282.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/6402
